+++ R/KenBoot.R
@@ -10,7 +10,9 @@
 #' @param proportion proportion to sub-sample, within [0,1[. WIP
 #' @param index Type of desired association index. At the molment only focal method implemented. WIP
 #' @param mode Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.
-#' @param output desired output of the bootstrap: edge lists, adjacency matrices, association index matrices, or igraph graphs
+#' @param output desired output of the bootstrap: edge lists, adjacency matrices, association index matrices, igraph graphs, or network metrics
+#' @param FUN if output is "metric", indicate which function to use to calculate said metric.
+#' @param ... if output is "metric", arguments for function FUN.
 #'
 #' @return according to output, returns a list of edge lists, adjacency matrices, association index matrices, or igraph graphs
 #' @export
@@ -20,7 +22,8 @@
 KenBoot<- function(Adj,effort,type = c("total","yab"),boot=100,replacement=TRUE,proportion=1.0,
                    index=c("joint","sri","hw","tw","sqrt","socaff","both","focal"),
                    mode = c("directed", "undirected", "max","min", "upper", "lower", "plus"),
-                   output=c("adjacency","edge","assoc","graph")){
+                   output=c("adjacency","edge","assoc","graph","metric"),
+                   FUN=NULL, ...){
 
   if(length(index)>1) {index<- "sri"}
   if(length(type)>1) {type<- "total"}
@@ -53,7 +56,7 @@ KenBoot<- function(Adj,effort,type = c("total","yab"),boot=100,replacement=TRUE,
 
   edge<- KenNet::adj.to.edge(Adj,mode = "directed")
 
-  if(output %in% c("edge","adjacency","assoc","graph")){
+  if(output %in% c("edge","adjacency","assoc","graph","metric")){
     e.boot<- lapply(1:boot,function(b) {
       if(replacement)
         edge[sample(1:nrow(edge),nrow(edge),replace = TRUE)]
@@ -61,21 +64,28 @@ KenBoot<- function(Adj,effort,type = c("total","yab"),boot=100,replacement=TRUE,
         edge[sample(1:nrow(edge),round(nrow(edge)*proportion,0),replace = FALSE)]
     }
     )
-    if(output %in% c("adjacency","assoc","graph")){
+    if(output %in% c("adjacency","assoc","graph","metric")){
       a.boot<- lapply(e.boot,function(edge) {
         KenNet::edge.to.adj(edge,mode = "directed")
       }
       )
-      if(output %in% c("assoc","graph")){
+      if(output %in% c("assoc","graph","metric")){
         assoc.boot<- lapply(a.boot,function(Adj) {
           KenNet::adj.to.assoc(Adj,effort = diag(Adj),index = "focal",type = "diag")
         }
         )
-        if(output %in% c("graph")){
+        if(output %in% c("graph","metric")){
           graph.boot<- lapply(assoc.boot,function(assoc) {
             igraph::graph_from_adjacency_matrix(assoc,mode = "plus",weighted = TRUE)
           }
           )
+          if(output %in% c("metric")){
+            fun.args<- list(...)
+            netmet.boot<- lapply(graph.boot,function(graph) {
+              KenNet::metric.boot(graph,FUN = FUN,...=fun.args)
+            }
+            )
+          }
         }
       }
     }
@@ -86,6 +96,7 @@ KenBoot<- function(Adj,effort,type = c("total","yab"),boot=100,replacement=TRUE,
           "adjacency" = a.boot,
           "assoc" = assoc.boot,
           "graph" = graph.boot,
+          "metric" = netmet.boot,
           stop("output type not recognized.")
   )
 }